Add initial end-to-end parsing test based on small examples
