+++ pymidi/chunks_test.py
@@ -1,0 +1,16 @@
+import unittest
+from pymidi.chunks import parse_chunks
+
+FORMAT_0_EXAMPLE = "data/format_0_example_1.mid"
+FORMAT_1_EXAMPLE = "data/format_1_example_1.mid"
+
+
+class ChunksTest(unittest.TestCase):
+
+    def test_parsing_format_0_file_results_in_header_and_track_chunk(self):
+        with open(FORMAT_0_EXAMPLE, "rb") as f:
+            chunks = parse_chunks(f)
+
+        self.assertEqual(len(chunks), 2)
+        self.assertEqual(chunks[0]["type"], "header")
+        self.assertEqual(chunks[1]["type"], "track")
